+++ src/main/java/com/zerobase/foodlier/module/member/member/client/OAuthApiClient.java
@@ -1,7 +1,7 @@
-package com.zerobase.foodlier.common.socialLogin.client;
+package com.zerobase.foodlier.module.member.member.client;
 
-import com.zerobase.foodlier.common.socialLogin.dto.OAuthInfoResponse;
-import com.zerobase.foodlier.common.socialLogin.dto.OAuthLoginParams;
+import com.zerobase.foodlier.module.member.member.social.dto.OAuthInfoResponse;
+import com.zerobase.foodlier.module.member.member.social.dto.OAuthLoginParams;
 import com.zerobase.foodlier.module.member.member.type.RegistrationType;
 
 public interface OAuthApiClient {
