+++ src/main/java/com/zerobase/foodlier/module/member/member/social/dto/NaverInfoResponse.java
@@ -1,44 +1,38 @@
-package com.zerobase.foodlier.common.socialLogin.dto;
+package com.zerobase.foodlier.module.member.member.social.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.zerobase.foodlier.module.member.member.type.RegistrationType;
 import lombok.Getter;
 
-import static com.zerobase.foodlier.module.member.member.type.RegistrationType.*;
+import static com.zerobase.foodlier.module.member.member.type.RegistrationType.NAVER;
 
 @Getter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class KakaoInfoResponse implements OAuthInfoResponse {
+public class NaverInfoResponse implements OAuthInfoResponse {
 
-    @JsonProperty("kakao_account")
-    private KakaoAccount kakaoAccount;
+    @JsonProperty("response")
+    private Response response;
 
     @Getter
     @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoAccount {
-        private KakaoProfile profile;
+    static class Response {
         private String email;
-    }
-
-    @Getter
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoProfile {
         private String nickname;
     }
 
     @Override
     public String getEmail() {
-        return kakaoAccount.email;
+        return response.email;
     }
 
     @Override
     public String getNickname() {
-        return kakaoAccount.profile.nickname;
+        return response.nickname;
     }
 
     @Override
     public RegistrationType getRegistrationType() {
-        return KAKAO;
+        return NAVER;
     }
 }
