+++ src/main/java/com/zerobase/foodlier/module/member/member/social/dto/NaverLoginParams.java
@@ -1,4 +1,4 @@
-package com.zerobase.foodlier.common.socialLogin.dto;
+package com.zerobase.foodlier.module.member.member.social.dto;
 
 import com.zerobase.foodlier.module.member.member.type.RegistrationType;
 import lombok.Getter;
@@ -6,22 +6,24 @@ import lombok.NoArgsConstructor;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import static com.zerobase.foodlier.module.member.member.type.RegistrationType.KAKAO;
+import static com.zerobase.foodlier.module.member.member.type.RegistrationType.NAVER;
 
 @Getter
 @NoArgsConstructor
-public class KakaoLoginParams implements OAuthLoginParams {
+public class NaverLoginParams implements OAuthLoginParams {
     private String authorizationCode;
+    private String state;
 
     @Override
     public RegistrationType registrationType() {
-        return KAKAO;
+        return NAVER;
     }
 
     @Override
     public MultiValueMap<String, String> makeBody() {
         MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
         body.add("code", authorizationCode);
+        body.add("state", state);
         return body;
     }
 }
