+++ src/main/java/com/zerobase/foodlier/module/member/member/social/dto/SocialLoginResponse.java
@@ -1,4 +1,4 @@
-package com.zerobase.foodlier.common.socialLogin.dto;
+package com.zerobase.foodlier.module.member.member.social.dto;
 
 import com.zerobase.foodlier.common.security.provider.dto.TokenDto;
 import com.zerobase.foodlier.module.member.member.type.RegistrationType;
